SparkViewEngine.ActiveViewFolder() is now an extension method
